requirements.txt
google-genai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
boto3>=1.28.0
pytest>=8.0.0
